requirements-raspi.txt
SpeechRecognition
requests
openai
python-dotenv
pyinstaller